clazy: msvc casts

+++ Software/src/SoundVisualizer.cpp
@@ -137,7 +137,7 @@ DECLARE_VISUALIZER(TwinPeaks, "Twin Peaks",
 public:
 private:
 	float m_previousPeak{ 0.0f };
-	size_t m_prevThresholdLed{ 0 };
+	unsigned int m_prevThresholdLed{ 0 };
 	double m_speedCoef = 1.0;
 	const uint8_t FadeOutSpeed = 12;
 );
@@ -145,12 +145,12 @@ private:
 bool TwinPeaksSoundVisualizer::visualize(const float* const fftData, const size_t fftSize, QList<QRgb>& colors)
 {
 	bool changed = false;
-	const size_t middleLed = std::floor(colors.size() / 2);
+	const unsigned int middleLed = std::floor(colors.size() / 2);
 
 	// most sensitive Hz range for humans
 	// this assumes 44100Hz sample rate
-	const size_t optimalHzMin = 1950 / (44100 / 2 / fftSize); // 2kHz
-	const size_t optimalHzMax = 5050 / (44100 / 2 / fftSize); // 5kHz
+	const unsigned int optimalHzMin = 1950 / (44100 / 2 / (unsigned int)fftSize); // 2kHz
+	const unsigned int optimalHzMax = 5050 / (44100 / 2 / (unsigned int)fftSize); // 5kHz
 
 	float currentPeak = 0.0f;
 	for (size_t i = 0; i < fftSize; ++i) {
@@ -165,7 +165,7 @@ bool TwinPeaksSoundVisualizer::visualize(const float* const fftData, const size_
 	else
 		m_previousPeak = std::max(0.0f, m_previousPeak - (fftSize * 0.000005f)); // lower the stale peak so it doesn't persist forever
 
-	const size_t thresholdLed = m_previousPeak != 0.0f ? middleLed * (currentPeak / m_previousPeak) : 0;
+	const unsigned int thresholdLed = m_previousPeak != 0.0f ? middleLed * (currentPeak / m_previousPeak) : 0;
 
 	// if leds move a lot with x% amplitude, increase fade speed
 	if (std::abs((int)(thresholdLed - m_prevThresholdLed)) > middleLed * 0.2)
@@ -173,8 +173,8 @@ bool TwinPeaksSoundVisualizer::visualize(const float* const fftData, const size_
 	else // reset on slow down
 		m_speedCoef = 1.0;// std::max(1.0, speedCoef - 2.0);
 
-	for (int idxA = 0; idxA < middleLed; ++idxA) {
-		const int idxB = colors.size() - 1 - idxA;
+	for (unsigned int idxA = 0; idxA < middleLed; ++idxA) {
+		const unsigned int idxB = colors.size() - 1 - idxA;
 		QRgb color = 0;
 		if (idxA < thresholdLed) {
 			QColor from = m_isLiquidMode ? m_generator.current() : m_minColor;
